requirements.txt
streamlit==1.28.1
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.1
transformers==4.36.2
torch==2.1.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pypdf==3.17.1
huggingface-hub==0.19.1